pyproject.toml
[project]
name = "dodentai"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "django-cors-headers>=4.4.0",
  "django-ninja-jwt>=5.3.3",
  "django-ninja>=1.3.0",
  "django-taggit>=6.1.0",
  "django>=5.1.1",
  "email-validator>=2.2.0",
  "environs>=11.0.0",
  "psycopg[binary]>=3.2.2",
  "redis[hiredis]>=5.0.8",
  "uvicorn>=0.30.6",
]

[tool.uv]
dev-dependencies = [
    "commitizen>=3.29.1",
    "djhtml>=3.0.6",
    "environs>=11.0.0",
    "faker>=30.3.0",
    "ruff>=0.6.9",
]
